Added unit tests

+++ src/main/java/com/example/movieapp/repository/MovieRepository.java
@@ -1,11 +1,12 @@
 package com.example.movieapp.repository;
 
 import com.example.movieapp.model.Movie;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface MovieRepository extends CrudRepository<Movie, Integer>
 {
-    int countByWatchedEquals(boolean watched);
+    long countByWatchedEquals(boolean watched);
 }
